face_recognition_api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
pydantic[email]>=2.4.2
pyodbc>=4.0.39
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.0.0